sd-netlink: various cleanups
- use uint8_t, uint16_t, and so on, rather than unsigned char, unsigned
  short, and so on, respectively,
- rename output parameters to ret or ret_xyz,
- add several missing assertions.

+++ src/network/networkd-nexthop.h
@@ -29,7 +29,7 @@ typedef struct NextHop {
         /* struct nhmsg */
         int family;
         uint8_t protocol;
-        uint8_t flags;
+        uint32_t flags;
 
         /* attributes */
         uint32_t id; /* NHA_ID */
